[install-sdk.sh] adding instructions for XLC install
The XLC Community Edition packages do not install simply by 'install
xlc'. They require extra switches (for yum) and the community license package.
Besides, a configure script needs to be run in order to set things up.

This patch adds the display of these commands at the end of the runtime
of install-sdk.sh, and also describes these modifications in the README
file.

Change-Id: I8c2872be32356a17de316313f8dc3fd2c5237133
Signed-off-by: Daniel Kreling <dbkreling@br.ibm.com>